+++ setup_fp_system_conda_python_hat.sh
@@ -2,19 +2,19 @@
 
 #-----------------------------------------------------------------------------------------
 # Script information
-script_name='FP ENVIRONMENT - PYTHON3 LIBRARIES FOR HMC PACKAGE'
-script_version="1.6.2"
-script_date='2021/06/28'
+script_name='FP ENVIRONMENT - PYTHON3 LIBRARIES FOR HAT PACKAGE - CONDA'
+script_version="1.6.3"
+script_date='2022/05/23'
 
 # Define file reference path according with https link(s) --> https://repo.anaconda.com/miniconda/
 fp_env_file_miniconda='https://repo.continuum.io/miniconda/Miniconda3-py37_4.8.2-Linux-x86_64.sh'
 
 # Argument(s) default definition(s)
-fp_env_folder_root_default=$HOME/fp_virtualenv_python3
-fp_env_file_reference_default='fp_virtualenv_python3_hmc_settings'
-fp_env_folder_libraries_default='fp_virtualenv_python3_hmc_libraries'
+fp_env_folder_root_default=$HOME/fp_system_conda
+fp_env_file_reference_default='fp_system_conda_python3_hat_settings'
+fp_env_folder_libraries_default='fp_system_conda_python3_hat_libraries'
 
-fp_env_file_requirements_default='requirements_fp_env_python_hmc.yaml'
+fp_env_file_requirements_default='requirements_fp_system_conda_python3_hat.yaml'
 
 # Examples of generic command-line:
 # conda create --yes --name $fp_env_folder_libraries numpy scipy pip python=3